renamed Switch to ToggleSwitch to avoid conflict
