+++ src/components/switch.css
@@ -2,9 +2,9 @@
     height: 0;
     width: 0;
     visibility: hidden;
-}
-
-.react-switch-label {
+  }
+  
+  .react-switch-label {
     display: flex;
     align-items: center;
     justify-content: space-between;
@@ -15,9 +15,9 @@
     border-radius: 100px;
     position: relative;
     transition: background-color .2s;
-}
-
-.react-switch-label .react-switch-button {
+  }
+  
+  .react-switch-label .react-switch-button {
     content: '';
     position: absolute;
     top: 2px;
@@ -28,13 +28,13 @@
     transition: 0.2s;
     background: #fff;
     box-shadow: 0 0 2px 0 rgba(10, 10, 10, 0.29);
-}
-
-.react-switch-checkbox + .react-switch-label .react-switch-button {
+  }
+  
+  .react-switch-checkbox:checked + .react-switch-label .react-switch-button {
     left: calc(100% - 2px);
-    transform: translate(-100%);
-}
-
-.react-switch-label:active .react-switch-button {
+    transform: translateX(-100%);
+  }
+  
+  .react-switch-label:active .react-switch-button {
     width: 60px;
-}
+  }
